Doc comments for query.
