+++ src/Microsoft.EntityFrameworkCore.SqlServer/Query/Internal/SqlServerQueryCompilationContextFactory.cs
@@ -5,7 +5,7 @@ using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore.Metadata;
-using Microsoft.EntityFrameworkCore.Query.ExpressionVisitors;
+using Microsoft.EntityFrameworkCore.Query.ExpressionVisitors.Internal;
 using Microsoft.EntityFrameworkCore.Utilities;
 using Remotion.Linq.Parsing.Structure.NodeTypeProviders;
 
@@ -32,23 +32,23 @@ namespace Microsoft.EntityFrameworkCore.Query.Internal
 
         public override QueryCompilationContext Create(bool async)
             => async
-                ? new SqlServerQueryCompilationContext(
-                    Model,
-                    (ISensitiveDataLogger)Logger,
-                    EntityQueryModelVisitorFactory,
-                    RequiresMaterializationExpressionVisitorFactory,
-                    new AsyncLinqOperatorProvider(),
-                    new AsyncQueryMethodProvider(),
-                    ContextType,
-                    TrackQueryResults)
-                : new SqlServerQueryCompilationContext(
-                    Model,
-                    (ISensitiveDataLogger)Logger,
-                    EntityQueryModelVisitorFactory,
-                    RequiresMaterializationExpressionVisitorFactory,
-                    new LinqOperatorProvider(),
-                    new QueryMethodProvider(),
-                    ContextType,
-                    TrackQueryResults);
+                   ? new SqlServerQueryCompilationContext(
+                       Model,
+                       (ISensitiveDataLogger)Logger,
+                       EntityQueryModelVisitorFactory,
+                       RequiresMaterializationExpressionVisitorFactory,
+                       new AsyncLinqOperatorProvider(),
+                       new AsyncQueryMethodProvider(),
+                       ContextType,
+                       TrackQueryResults)
+                   : new SqlServerQueryCompilationContext(
+                       Model,
+                       (ISensitiveDataLogger)Logger,
+                       EntityQueryModelVisitorFactory,
+                       RequiresMaterializationExpressionVisitorFactory,
+                       new LinqOperatorProvider(),
+                       new QueryMethodProvider(),
+                       ContextType,
+                       TrackQueryResults);
     }
 }
