+++ src/Microsoft.EntityFrameworkCore/Microsoft.EntityFrameworkCore.csproj
@@ -372,12 +372,12 @@
     <Compile Include="Query\EntityQueryModelVisitor.cs" />
     <Compile Include="Query\EntityQueryModelVisitorFactory.cs" />
     <Compile Include="Query\ExpressionVisitors\EntityQueryableExpressionVisitor.cs" />
-    <Compile Include="Query\ExpressionVisitors\ExpressionTransformingQueryModelVisitor.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\ExpressionTransformingQueryModelVisitor.cs" />
     <Compile Include="Query\ExpressionVisitors\ExpressionVisitorBase.cs" />
     <Compile Include="Query\ExpressionVisitors\IEntityQueryableExpressionVisitorFactory.cs" />
-    <Compile Include="Query\ExpressionVisitors\IEntityResultFindingExpressionVisitorFactory.cs" />
-    <Compile Include="Query\ExpressionVisitors\IMemberAccessBindingExpressionVisitorFactory.cs" />
-    <Compile Include="Query\ExpressionVisitors\INavigationRewritingExpressionVisitorFactory.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\IEntityResultFindingExpressionVisitorFactory.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\IMemberAccessBindingExpressionVisitorFactory.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\INavigationRewritingExpressionVisitorFactory.cs" />
     <Compile Include="Query\ExpressionVisitors\Internal\CompositeKey.cs" />
     <Compile Include="Query\ExpressionVisitors\Internal\EntityEqualityRewritingExpressionVisitor.cs" />
     <Compile Include="Query\ExpressionVisitors\Internal\DefaultQueryExpressionVisitor.cs" />
@@ -397,12 +397,12 @@
     <Compile Include="Query\ExpressionVisitors\Internal\RequiresMaterializationExpressionVisitorFactory.cs" />
     <Compile Include="Query\ExpressionVisitors\Internal\SubQueryMemberPushDownExpressionVisitor.cs" />
     <Compile Include="Query\ExpressionVisitors\Internal\TaskBlockingExpressionVisitor.cs" />
-    <Compile Include="Query\ExpressionVisitors\IOrderingExpressionVisitorFactory.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\IOrderingExpressionVisitorFactory.cs" />
     <Compile Include="Query\ExpressionVisitors\IProjectionExpressionVisitorFactory.cs" />
-    <Compile Include="Query\ExpressionVisitors\IQuerySourceTracingExpressionVisitorFactory.cs" />
-    <Compile Include="Query\ExpressionVisitors\IRequiresMaterializationExpressionVisitorFactory.cs" />
-    <Compile Include="Query\ExpressionVisitors\ISubQueryMemberPushDownExpressionVisitor.cs" />
-    <Compile Include="Query\ExpressionVisitors\ITaskBlockingExpressionVisitor.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\IQuerySourceTracingExpressionVisitorFactory.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\IRequiresMaterializationExpressionVisitorFactory.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\ISubQueryMemberPushDownExpressionVisitor.cs" />
+    <Compile Include="Query\ExpressionVisitors\Internal\ITaskBlockingExpressionVisitor.cs" />
     <Compile Include="Query\ExpressionVisitors\ProjectionExpressionVisitor.cs" />
     <Compile Include="Query\ICompiledQueryCacheKeyGenerator.cs" />
     <Compile Include="Query\IEntityQueryModelVisitorFactory.cs" />
@@ -450,7 +450,7 @@
     <Compile Include="Query\QueryCompilationContext.cs" />
     <Compile Include="Query\QueryContext.cs" />
     <Compile Include="Query\QueryContextFactory.cs" />
-    <Compile Include="Query\QuerySourceExtensions.cs" />
+    <Compile Include="Query\Internal\QuerySourceExtensions.cs" />
     <Compile Include="Query\ResultOperatorHandler.cs" />
     <Compile Include="Query\ResultOperators\Internal\IncludeExpressionNode.cs" />
     <Compile Include="Query\ResultOperators\Internal\IncludeResultOperator.cs" />
