+++ src/Microsoft.EntityFrameworkCore/Query/ExpressionVisitors/ExpressionVisitorBase.cs
@@ -8,13 +8,23 @@ using Remotion.Linq.Parsing;
 
 namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors
 {
+    /// <summary>
+    ///     A base expression visitor that ignores Block expressions.
+    /// </summary>
     [DebuggerStepThrough]
     public abstract class ExpressionVisitorBase : RelinqExpressionVisitor
     {
+        /// <summary>
+        ///     Visits the given node.
+        /// </summary>
+        /// <param name="node"> The expression to visit. </param>
+        /// <returns>
+        ///     An Expression.
+        /// </returns>
         public override Expression Visit([CanBeNull] Expression node)
             => node == null
                || node.NodeType == ExpressionType.Block
-                ? node
-                : base.Visit(node);
+                   ? node
+                   : base.Visit(node);
     }
 }
