--- src/Microsoft.EntityFrameworkCore/Query/ExpressionVisitors/IMemberAccessBindingExpressionVisitorFactory.cs
@@ -1,17 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Linq.Expressions;
-using JetBrains.Annotations;
-using Remotion.Linq.Clauses;
-
-namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors
-{
-    public interface IMemberAccessBindingExpressionVisitorFactory
-    {
-        ExpressionVisitor Create(
-            [NotNull] QuerySourceMapping querySourceMapping,
-            [NotNull] EntityQueryModelVisitor queryModelVisitor,
-            bool inProjection);
-    }
-}
