--- src/Microsoft.EntityFrameworkCore/Query/ExpressionVisitors/ISubQueryMemberPushDownExpressionVisitor.cs
@@ -1,13 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Linq.Expressions;
-using JetBrains.Annotations;
-
-namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors
-{
-    public interface ISubQueryMemberPushDownExpressionVisitor
-    {
-        Expression Visit([CanBeNull] Expression expression);
-    }
-}
