+++ src/Microsoft.EntityFrameworkCore/Query/ExpressionVisitors/Internal/IMemberAccessBindingExpressionVisitorFactory.cs
@@ -1,0 +1,25 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Linq.Expressions;
+using JetBrains.Annotations;
+using Remotion.Linq.Clauses;
+
+namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors.Internal
+{
+    /// <summary>
+    ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
+    ///     directly from your code. This API may change or be removed in future releases.
+    /// </summary>
+    public interface IMemberAccessBindingExpressionVisitorFactory
+    {
+        /// <summary>
+        ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
+        ///     directly from your code. This API may change or be removed in future releases.
+        /// </summary>
+        ExpressionVisitor Create(
+            [NotNull] QuerySourceMapping querySourceMapping,
+            [NotNull] EntityQueryModelVisitor queryModelVisitor,
+            bool inProjection);
+    }
+}
