+++ src/Microsoft.EntityFrameworkCore/Query/ExpressionVisitors/Internal/IRequiresMaterializationExpressionVisitorFactory.cs
@@ -1,0 +1,20 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using JetBrains.Annotations;
+
+namespace Microsoft.EntityFrameworkCore.Query.ExpressionVisitors.Internal
+{
+    /// <summary>
+    ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
+    ///     directly from your code. This API may change or be removed in future releases.
+    /// </summary>
+    public interface IRequiresMaterializationExpressionVisitorFactory
+    {
+        /// <summary>
+        ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
+        ///     directly from your code. This API may change or be removed in future releases.
+        /// </summary>
+        RequiresMaterializationExpressionVisitor Create([NotNull] EntityQueryModelVisitor queryModelVisitor);
+    }
+}
