+++ src/Microsoft.EntityFrameworkCore/Query/ICompiledQueryCacheKeyGenerator.cs
@@ -6,8 +6,17 @@ using JetBrains.Annotations;
 
 namespace Microsoft.EntityFrameworkCore.Query
 {
+    /// <summary>
+    ///     A cache key generator for the compiled query cache.
+    /// </summary>
     public interface ICompiledQueryCacheKeyGenerator
     {
+        /// <summary>
+        ///     Generates a cache key.
+        /// </summary>
+        /// <param name="query"> The query to generate a cache key for. </param>
+        /// <param name="async"> True if the query will be executed asynchronously. </param>
+        /// <returns> An object representing a query cache key. </returns>
         object GenerateCacheKey([NotNull] Expression query, bool async);
     }
 }
