+++ src/Microsoft.EntityFrameworkCore/Query/IEntityQueryModelVisitorFactory.cs
@@ -5,8 +5,19 @@ using JetBrains.Annotations;
 
 namespace Microsoft.EntityFrameworkCore.Query
 {
+    /// <summary>
+    ///     A factory for creating EntityQueryModelVisitors.
+    /// </summary>
     public interface IEntityQueryModelVisitorFactory
     {
+        /// <summary>
+        ///     Creates a new EntityQueryModelVisitor.
+        /// </summary>
+        /// <param name="queryCompilationContext"> Context for the query compilation. </param>
+        /// <param name="parentEntityQueryModelVisitor"> The parent entity query model visitor. </param>
+        /// <returns>
+        ///     An EntityQueryModelVisitor instance.
+        /// </returns>
         EntityQueryModelVisitor Create(
             [NotNull] QueryCompilationContext queryCompilationContext,
             [CanBeNull] EntityQueryModelVisitor parentEntityQueryModelVisitor);
