+++ src/Microsoft.EntityFrameworkCore/Query/IIncludableQueryable.cs
@@ -5,6 +5,11 @@ using System.Linq;
 
 namespace Microsoft.EntityFrameworkCore.Query
 {
+    /// <summary>
+    ///     Supports queryable Include/ThenInclude chaining operators.
+    /// </summary>
+    /// <typeparam name="TEntity"> The entity type. </typeparam>
+    /// <typeparam name="TProperty"> The property type. </typeparam>
     // ReSharper disable once UnusedTypeParameter
     public interface IIncludableQueryable<out TEntity, out TProperty> : IQueryable<TEntity>
     {
