+++ src/Microsoft.EntityFrameworkCore/Query/IQueryCompilationContextFactory.cs
@@ -3,8 +3,18 @@
 
 namespace Microsoft.EntityFrameworkCore.Query
 {
+    /// <summary>
+    ///     Factory for <see cref="QueryCompilationContext" /> instances.
+    /// </summary>
     public interface IQueryCompilationContextFactory
     {
+        /// <summary>
+        ///     Creates a new QueryCompilationContext.
+        /// </summary>
+        /// <param name="async"> true if the query will be executed asynchronously. </param>
+        /// <returns>
+        ///     A <see cref="QueryCompilationContext" /> instance.
+        /// </returns>
         QueryCompilationContext Create(bool async);
     }
 }
