+++ src/Microsoft.EntityFrameworkCore/Query/IResultOperatorHandler.cs
@@ -8,8 +8,20 @@ using Remotion.Linq.Clauses;
 
 namespace Microsoft.EntityFrameworkCore.Query
 {
+    /// <summary>
+    ///     A handler for <see cref="ResultOperatorBase" /> instances.
+    /// </summary>
     public interface IResultOperatorHandler
     {
+        /// <summary>
+        ///     Handles a result operator.
+        /// </summary>
+        /// <param name="entityQueryModelVisitor"> The entity query model visitor. </param>
+        /// <param name="resultOperator"> The result operator. </param>
+        /// <param name="queryModel"> The query model. </param>
+        /// <returns>
+        ///     A compiled query expression fragment representing the result operator.
+        /// </returns>
         Expression HandleResultOperator(
             [NotNull] EntityQueryModelVisitor entityQueryModelVisitor,
             [NotNull] ResultOperatorBase resultOperator,
