+++ src/Microsoft.EntityFrameworkCore/Query/IncludeSpecification.cs
@@ -11,8 +11,16 @@ using Remotion.Linq.Clauses;
 
 namespace Microsoft.EntityFrameworkCore.Query
 {
+    /// <summary>
+    ///     Represents a single query include operation.
+    /// </summary>
     public class IncludeSpecification
     {
+        /// <summary>
+        ///     Initializes a new instance of the Microsoft.EntityFrameworkCore.Query.IncludeSpecification class.
+        /// </summary>
+        /// <param name="querySource"> The query source. </param>
+        /// <param name="navigationPath"> The set of navigation properties to be included. </param>
         public IncludeSpecification(
             [NotNull] IQuerySource querySource,
             [NotNull] IReadOnlyList<INavigation> navigationPath)
@@ -24,11 +32,36 @@ namespace Microsoft.EntityFrameworkCore.Query
             NavigationPath = navigationPath;
         }
 
+        /// <summary>
+        ///     Gets the query source.
+        /// </summary>
+        /// <value>
+        ///     The query source.
+        /// </value>
         public virtual IQuerySource QuerySource { get; }
+
+        /// <summary>
+        ///     Gets the set of navigation properties to be included.
+        /// </summary>
+        /// <value>
+        ///     The set of navigation properties to be included.
+        /// </value>
         public virtual IReadOnlyList<INavigation> NavigationPath { get; }
 
+        /// <summary>
+        ///     Gets or sets a value indicating whether this object is an enumerable target.
+        /// </summary>
+        /// <value>
+        ///     True if this object is an enumerable target, false if not.
+        /// </value>
         public virtual bool IsEnumerableTarget { get; set; }
 
+        /// <summary>
+        ///     Convert this object into a string representation.
+        /// </summary>
+        /// <returns>
+        ///     A string that represents this object.
+        /// </returns>
         public override string ToString()
         {
             return $"{QuerySource.ItemName}.{NavigationPath.Select(n => n.Name).Join(".")}";
