+++ src/Microsoft.EntityFrameworkCore/Query/ResultOperators/IQueryAnnotation.cs
@@ -7,9 +7,25 @@ using Remotion.Linq.Clauses;
 
 namespace Microsoft.EntityFrameworkCore.Query.ResultOperators
 {
+    /// <summary>
+    ///     Represents an annotation on a query.
+    /// </summary>
     public interface IQueryAnnotation
     {
+        /// <summary>
+        ///     Gets the query source.
+        /// </summary>
+        /// <value>
+        ///     The query source.
+        /// </value>
         IQuerySource QuerySource { get; [param: NotNull] set; }
+
+        /// <summary>
+        ///     Gets the query model.
+        /// </summary>
+        /// <value>
+        ///     The query model.
+        /// </value>
         QueryModel QueryModel { get; [param: NotNull] set; }
     }
 }
